Moved BlockLocParser and TransactionLocParser to parser folder.
