+++ tests/unit/utils/BlockLocParser.spec.ts
@@ -26,7 +26,7 @@ import {SAMPLE_BLOCK} from "../Mocks";
 import MockAdapter from "axios-mock-adapter";
 import axios from "axios";
 import {CacheUtils} from "@/utils/cache/CacheUtils";
-import {BlockLocParser} from "@/utils/BlockLocParser";
+import {BlockLocParser} from "@/utils/parser/BlockLocParser";
 import {PathParam} from "@/utils/PathParam";
 
 describe("BlockLocParser.ts", () => {
